Modifed MatDeepRep function and created GetFeatures Script
MatDeepRep now returns 4 variables, each containing the features for
positive and negative features for training and testing, with the
purpose of potentially training both an SVM and a Neural Network.
The script GetFeatures is planned to automatise the process of running
the MatDeepRep function on each Model/Category combination and saving
the extracted features on a txt file for further use.